# docs/openapi-pre-decomposition.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stitchd Feature Flag API",
    "description": "REST API for managing feature flags, segments, and evaluation.",
    "contact": {
      "name": "Stitchd"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/environments/{env_id}/event-definitions": {
      "get": {
        "tags": [
          "event-definitions"
        ],
        "summary": "`GET /v1/environments/{env_id}/event-definitions` — List active event definitions.",
        "description": "# Errors\nReturns 500 if the repository query fails.",
        "operationId": "list_event_definitions",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of event definitions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventDefinitionResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "post": {
        "tags": [
          "event-definitions"
        ],
        "summary": "`POST /v1/environments/{env_id}/event-definitions` — Register an event definition.",
        "description": "# Errors\nReturns 409 if a definition with the same key already exists in the environment.",
        "operationId": "create_event_definition",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDefinitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Event definition registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDefinitionResponse"
                }
              }
            }
          },
          "409": {
            "description": "Key already exists in this environment"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/environments/{env_id}/event-definitions/{key}": {
      "delete": {
        "tags": [
          "event-definitions"
        ],
        "summary": "`DELETE /v1/environments/{env_id}/event-definitions/{key}` — Soft-delete by key.",
        "description": "# Errors\nReturns 404 if no active definition with that key exists.",
        "operationId": "delete_event_definition",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "in": "path",
            "description": "Event definition key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Event definition deleted"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/environments/{env_id}/events": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "`POST /v1/environments/{env_id}/events` — Ingest a single event.",
        "description": "Validates the `metric_key` against registered definitions and checks the\nvalue type. Returns 202 immediately; the ClickHouse write is async.\n\n# Errors\n\nReturns [`ApiError`] if validation fails or an internal error occurs.",
        "operationId": "ingest_single_event",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Event accepted for ingestion"
          },
          "401": {
            "description": "Missing or invalid SDK key"
          },
          "422": {
            "description": "Unknown key or type mismatch"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "security": [
          {
            "sdk_key": []
          }
        ]
      }
    },
    "/v1/environments/{env_id}/events/batch": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "`POST /v1/environments/{env_id}/events/batch` — Ingest up to 500 events.",
        "description": "All events are validated before any write is initiated. A single invalid\nevent causes the entire batch to be rejected with 422.\n\n# Errors\n\nReturns [`ApiError`] if the batch is too large, validation fails, or an internal error occurs.",
        "operationId": "ingest_batch_events",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchEventPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Batch accepted for ingestion"
          },
          "401": {
            "description": "Missing or invalid SDK key"
          },
          "422": {
            "description": "Batch too large, unknown key, or type mismatch"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "security": [
          {
            "sdk_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchEventPayload": {
        "type": "object",
        "description": "Payload for a batch event ingestion request (max 500 events).",
        "required": [
          "events"
        ],
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventPayload"
            },
            "description": "Events to ingest; maximum 500."
          }
        }
      },
...
      "CreateEventDefinitionRequest": {
        "type": "object",
        "description": "Request body for registering a new event definition.",
        "required": [
          "key",
          "value_type"
        ],
        "properties": {
          "key": {
            "type": "string",
            "description": "URL-safe string key, unique within the environment."
          },
          "value_type": {
            "$ref": "#/components/schemas/EventValueType",
            "description": "The metric value type this event will carry."
          }
        }
      },
...
      "EnvironmentId": {
        "type": "string",
        "format": "uuid"
      },
...
      "EventContext": {
        "type": "object",
        "description": "A single context dimension attached to an event (e.g. `user:u1`, `session:s99`).",
        "required": [
          "_type",
          "key"
        ],
        "properties": {
          "_type": {
            "type": "string",
            "description": "Dimension type (e.g. `\"user\"`, `\"session\"`)."
          },
          "key": {
            "type": "string",
            "description": "Dimension value (e.g. `\"u1\"`)."
          }
        }
      },
      "EventDefinitionId": {
        "type": "string",
        "format": "uuid"
      },
      "EventDefinitionResponse": {
        "type": "object",
        "description": "Response body for a single event definition.",
        "required": [
          "id",
          "environment_id",
          "key",
          "value_type",
          "created_at",
          "updated_at",
          "version"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 creation timestamp."
          },
          "environment_id": {
            "$ref": "#/components/schemas/EnvironmentId",
            "description": "The environment this definition belongs to."
          },
          "id": {
            "$ref": "#/components/schemas/EventDefinitionId",
            "description": "Unique identifier."
          },
          "key": {
            "type": "string",
            "description": "URL-safe string key, unique within the environment."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "ISO-8601 last-updated timestamp."
          },
          "value_type": {
            "$ref": "#/components/schemas/EventValueType",
            "description": "The metric value type this event carries."
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "description": "Optimistic-concurrency version counter."
          }
        }
      },
      "EventPayload": {
        "type": "object",
        "description": "Payload for a single event ingestion request.",
        "required": [
          "contexts",
          "metric_key",
          "value",
          "timestamp"
        ],
        "properties": {
          "contexts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventContext"
            },
            "description": "Context dimensions for this event (e.g. user, session)."
          },
          "metric_key": {
            "type": "string",
            "description": "Pre-registered metric key for this event."
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Client-supplied event timestamp."
          },
          "value": {
            "$ref": "#/components/schemas/EventValue",
            "description": "Typed metric value; must match the registered value type."
          }
        }
      },
      "EventValue": {
        "oneOf": [
          {
            "type": "object",
            "description": "Boolean metric.",
            "required": [
              "bool"
            ],
            "properties": {
              "bool": {
                "type": "boolean",
                "description": "Boolean metric."
              }
            }
          },
          {
            "type": "object",
            "description": "64-bit integer metric.",
            "required": [
              "int"
            ],
            "properties": {
              "int": {
                "type": "integer",
                "format": "int64",
                "description": "64-bit integer metric."
              }
            }
          },
          {
            "type": "object",
            "description": "64-bit floating-point metric.",
            "required": [
              "double"
            ],
            "properties": {
              "double": {
                "type": "number",
                "format": "double",
                "description": "64-bit floating-point metric."
              }
            }
          }
        ],
        "description": "The typed metric value carried by an event.\n\nExactly one variant must be present; the variant must match the registered\n`EventValueType` for the event's `metric_key` or the event is rejected (422)."
      },
      "EventValueType": {
        "type": "string",
        "description": "The metric value type an event definition accepts.",
        "enum": [
          "bool",
          "int",
          "double"
        ]
      },
...
    }
  }
}